+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Chalk Release Notes
 
-## On the `main` branch
+## 0.4.3
+
+**June 10, 2024**
 
 ### Fixes